% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplodata4ggplot.R
\name{haplodata4ggplot}
\alias{haplodata4ggplot}
\title{haplodata4ggplot}
\usage{
haplodata4ggplot(haploNet_data_object, method = "pegas")
}
\arguments{
\item{haploNet_data_object}{A list containing the haplotype network data
named network, and a dataframe named pieseas. The latter contains the
proportions of each sample or group of samples for each haplotype. This 
object can be generated using the function haploNet_data.

 @param method A string indicating the method to use for plotting the network.
Options are "pegas" for the pegas package or "fruchtermanreingold" for the
Fruchterman-Reingold algorithm.}
}
\value{
Returns a list with two data frames: nodes and connections.
The nodes data frame contains the x and y coordinates, size, labels, and
proportions of each haplotype. The connections data frame contains the
coordinates of the connections between nodes.
}
\description{
This function creates a list containing nodes and connections for a haplotype
network plot, suitable for use with ggplot2.
}
\details{
This function prepares the data for plotting haplotype networks, 
using either the pegas package or the Fruchterman-Reingold algorithm to
determine node positions. The nodes are represented as pie charts, with
sizes proportional to haplotype frequencies. The aim of this function is to
facilitate the creation of publication-ready haplotype network plots using
ggplot2, addressing common issues such as legend overlap with the network.
}
